+++ meteo-entity/src/main/java/no/api/meteo/entity/core/Location.java
@@ -18,7 +18,6 @@ package no.api.meteo.entity.core;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import net.sf.oval.constraint.NotNegative;
 import net.sf.oval.constraint.NotNull;
@@ -38,12 +37,12 @@ public class Location {
 
     @NotNegative
     @JsonProperty
-    private final Double altitude;
+    private final Integer altitude;
 
     @JsonCreator
     public Location(@JsonProperty("longitude") Double longitude,
                     @JsonProperty("latitude") Double latitude,
-                    @JsonProperty("altitude") Double altitude) {
+                    @JsonProperty("altitude") Integer altitude) {
         this.longitude = longitude;
         this.latitude = latitude;
         this.altitude = altitude;
